[ORC] Runtime work-in-progress. Take 3.
Adds a compiler-rt library (compiler-rt/lib/orc) that includes support for
cxa_atexit interposing; dlfcn.h function emulation; and a new runtime function,
__orc_rt_macho_run_program, for executing JIT'd programs by dlopening the
Main JITDylib, dlsym'ing main, then dlclosing the JITDylib. Emulation of the
dlfcn functions includes running static constructors and destructors.

The llvm-jitlink tool is updated to automatically load the runtime on Darwin and
use it to run target programs. The effect of this is that objects added to
llvm-jitlink (at least on Darwin) will now have their static constructors and
destructors executed, both in in-process mode and out-of-process mode.

Beyond Take 2:

Switches from ORC-RPC to the new Blob serialization system which is not tied to
llvm::Error, and should be able to more efficiently serialize small types.

Abandons direct calling between ORC and the runtime: The runtime is allowed to
be commpiled with a different ABI (e.g. breaking checks) so we should always
serialize when calling in to and out of it.

This patch reduces the ORC runtime's dependencies on LLVM headers to
llvm/ExecutionEngine/Shared/WrapperFunctionTypes.h and
llvm/ExeuctionEngien/Shared/MachOPlatformTypes.h, which in turn only
depend on llvm/Support/SwapByteOrder.h, llvm-c/LLVMOrcShared.h and system
headers.

Also aggregates a lot of code in orc-rt.