[ORC] Add EPCDynamicLibrarySearchGenerator::LoadWeak method.
LoadWeak takes a required Allow predicate and a path to a dylib that may or may
not be present. If there is a valid dylib at the path then the generator
behaves the same as if constructed with the Load method (i.e. the dylib is
loaded and symbols resolved within it). If there is no valid dylib at the path
then the generator resolves all symbols matching the Allow predicate to null.

This allows LoadWeak to match the behavior of ld64's -weak_library / -weak-lx
options.

Note that this does not affect the behavior of the llvm-jitlink tool's
-weak_library / -weak-lx options, which always treat the library as missing
even if the path points to a valid dylib.