[ORC] Generalize dispatch handler registration, add deregistration.
Previously ExecutionSession::registerJITDispatchHandlers supported associating
handlers with named symbols in JIT'd code. This commit generalizes the
interface to allow handlers to be associated with arbitrary addresses in the
JIT'd process, simplify registration of handlers for remote objects*.
Deregistration support is also added so that remote objects can be released
prior to the end of the session.

* When an object is created in the executor, an address associated with that
  particular object can be registered in the controller to receive calls from
  it.